Move cassandra driver version to 4.14.1 && leverage client configuration file
Use the latest version of the cassandra driver and make the adequate
code changes.
As the new client version offer a way to set the configuration through a
file, I took the opportunity to leverage it to simplify the plugin
configuration (just link to a full fledge configuration file should
suffice).